fix dtype error

+++ xturing/engines/llama_engine.py
@@ -258,7 +258,8 @@ class LlamaLoraInt4Engine(CausalLoraEngine):
         torch.nn.init.normal_ = saved_normal_
 
         # only training mm_projector
-        torch.set_default_dtype(torch.float)
+        # torch.set_default_dtype(torch.float) # when image_features = self.mm_projector(image_features): RuntimeError: expected scalar type Half but found Float
+        torch.set_default_dtype(torch.half)
            
 
         if pretrain_mm_mlp_adapter is not None:
